Add automatic binary jar file stubbing to lint's test infrastructure
When testing detectors which are accessing stubs of a particular
library, it's often useful to include those stubs as *binary* classes,
since there are differences between the PSI you get from a library jar
and from a local source file. As just a simple example, constants will
have been inlined.

Until now, creating binary stubs was cumbersome; you'd need to compile
them somewhere, and then copy in the gzip compressed, base64-encoded
class files.

This CL adds a new test DSL method which makes this simpler:

   mavenLibrary(artifact, *sourceFiles)

where you can specify the stubs as simple source files, and lint will
dynamically create class files on the fly, package these as a jar, and
place the jar in the test project instead.

The way this works is that lint will first parse the source file using
UAST, and then it visits the UAST and uses an ASM ClassWriter to write
a series of corresponding class files.

This CL also updates some existing tests which used base64gzipped stubs
over to using this new mechanism (but the real driver for this feature
is an upcoming detector for dependency analysis which will needs to test
a lot of library stubs.)

As part of this I discovered and fixed two bugs:

- The code to generate a method descriptor produced an incorrect
  descriptor when you call a constructor of a non-static innerclass.  In
  this case, the bytecode will have a synthetic argument (to the
  outerclass) as the first parameter, which was missing/put in the wrong
  place.

- The test mode support for testing source-only and bytecode-only
  versions wasn't working right for compiled test files packaged as jars
  (since the sources were placed inside the .jar file).

Test: Existing tests, and many new unit tests as well
Bug: Not filed

+++ lint/libs/lint-tests/src/main/java/com/android/tools/lint/checks/infrastructure/TestMode.kt
@@ -15,7 +15,6 @@
  */
 package com.android.tools.lint.checks.infrastructure
 
-import com.android.SdkConstants.DOT_JAR
 import com.android.SdkConstants.DOT_KT
 import com.android.SdkConstants.DOT_KTS
 import com.android.tools.lint.checks.infrastructure.TestMode.Companion.UI_INJECTION_HOST
@@ -144,33 +143,46 @@ open class TestMode(
             // Delete sources for any compiled files since when analyzing a project
             // you can only see the local sources.
             for (project in projects) {
-                for (file in project.files) {
+                // If we're testing a check that only analyzes bytecode, don't delete sources
+                // or bytecode
+                if (context.task.issues == null || context.task.issues.all { it.implementation.classOnly() }) {
+                    return
+                }
+
+                val projectFiles = project.files
+                for (file in projectFiles) {
                     // TODO: Consider whether I need to remove sources from within jars too?
                     // Check whether resolve finds them first.
 
-                    if (file !is CompiledSourceFile) {
+                    val sources: List<TestFile>
+                    val classFiles: List<String>
+                    if (file is BytecodeTestFile) {
+                        if (file.getSources().any { it in projectFiles }) {
+                            // Unusual, but see NotificationPermissionDetectorTest#testClassAndSourceFileUsage
+                            // where we're testing the combination of byte code and source code where the
+                            // byte code compiled source is aliasing an actual source which would get deleted.
+                            continue
+                        }
+                        sources = if (deleteSourceFiles) file.getSources() else emptyList()
+                        classFiles = if (deleteBinaryFiles) file.getGeneratedPaths() else emptyList()
+                    } else {
                         continue
                     }
 
-                    // If we're testing a check that only analyzes bytecode, don't delete sources
-                    // or bytecode
-                    if (context.task.issues.all { it.implementation.classOnly() }) {
-                        return
-                    }
-
-                    if (deleteSourceFiles && !file.targetRelativePath.endsWith(DOT_JAR)) {
+                    if (deleteSourceFiles) {
                         for (dir in context.projectFolders) {
-                            val source = File(dir, file.source.targetPath)
-                            if (source.exists()) {
-                                source.delete()
-                                break
+                            for (testFile in sources) {
+                                val source = File(dir, testFile.targetPath)
+                                if (source.exists()) {
+                                    source.delete()
+                                }
                             }
                         }
                     }
                     if (deleteBinaryFiles) {
-                        for (classTestFile in file.classFiles) {
+                        for (classTestFile in classFiles) {
                             for (dir in context.projectFolders) {
-                                val classFile = File(dir, classTestFile.targetPath)
+                                val classFile = File(dir, classTestFile)
                                 if (classFile.exists()) {
                                     classFile.delete()
                                     break
@@ -236,7 +248,7 @@ open class TestMode(
             override fun applies(context: TestModeContext): Boolean {
                 return context.projects.any {
                     it.files.any { file ->
-                        file is CompiledSourceFile && file.type == CompiledSourceFile.Type.SOURCE_AND_BYTECODE
+                        file is BytecodeTestFile && file.type == BytecodeTestFile.Type.SOURCE_AND_BYTECODE
                     }
                 }
             }
